// src/api/generated/model/applicationDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ApplicationDto {
  id?: string;
  projectId?: string;
  projectName?: string;
  userId?: string;
  userName?: string;
  userEmail?: string;
  status?: string;
  appliedAt?: string;
  decidedAt?: string;
  decidedById?: string;
  decidedByName?: string;
  message?: string;
}